[Refactor] フォロー機能のリファクタリング

+++ app/controllers/public/relationships_controller.rb
@@ -2,20 +2,14 @@ class Public::RelationshipsController < ApplicationController
   before_action :authenticate_user!
   before_action :ensure_guest_user
   before_action :deny_deactivated_user
+  before_action :set_user
 
   def create
-    if current_user.id == params[:user_id].to_i
-      flash[:notice] = '自分自身はフォローできません。'
-    else
+    if @user.id != current_user.id && @user.is_active # 自分自身と、退会済ユーザーはフォローできない
       # current_user.followers.newで「follower_user_idカラムにcurrent_user.idが入ったデータ」を作成する
       new_relationship = current_user.followers.new(followed_user_id: params[:user_id])
       new_relationship.save
-      # 重複フォローの防止
-      if new_relationship.errors.any?
-        flash[:notice] = '既にフォロー中です。'
-      end
     end
-    @user = User.find(params[:user_id]) # インスタンス変数で宣言するのはjQueryでの部分テンプレート呼出時に必要なため
     respond_to do |format|
       if params[:from_page] == 'users_show'
         format.js { render :create_at_users_show }
@@ -27,9 +21,11 @@ class Public::RelationshipsController < ApplicationController
 
   def destroy
     # current_user.followers.find_byで「follower_user_idカラムにcurrent_user.idが入ったデータ」にfind_byを実行する
-    destroy_relationship = current_user.followers.find_by(followed_user_id: params[:user_id])
-    destroy_relationship.destroy
-    @user = User.find(params[:user_id]) # インスタンス変数で宣言するのはjQueryでの部分テンプレート呼出時に必要なため
+    destroy_relationship = current_user.followers.find_by(followed_user_id: @user.id)
+    # destroy_relationshipがnilでないときのみdestroyを実行 別の画面でフォローを解除していた場合と、対象ユーザーが退会していた時のNoMethodErrorエラー回避
+    if destroy_relationship
+      destroy_relationship.destroy
+    end
     respond_to do |format|
       if params[:from_page] == 'users_show'
         format.js { render :destroy_at_users_show }
@@ -40,6 +36,10 @@ class Public::RelationshipsController < ApplicationController
   end
 
   private
+  def set_user
+    @user = User.find(params[:user_id]) # インスタンス変数で宣言するのはjQueryでの部分テンプレート呼出時に必要なため
+  end
+
   def ensure_guest_user
     if current_user.guest_user?
       redirect_to posts_path, notice: 'ゲストユーザーはフォロー機能を使用できません。'
